workers/requirements.txt
Pillow==12.1.1
ImageHash==4.3.2
numpy==2.4.3
